Rename mod `request`` to `release_artifacts`
